requirements.txt
python-dotenv
pyyaml
claude-agent-sdk
slack_sdk
